Promote findKeyValue to a generic utility function
