+++ ui/src/utils/findKeyValue.ts
@@ -1,0 +1,26 @@
+import { Maybe, StopRegistryKeyValues } from '../generated/graphql';
+
+type ElementWithKeyValues = {
+  readonly keyValues?: Maybe<Array<Maybe<StopRegistryKeyValues>>>;
+};
+
+export function findKeyValue(
+  element: ElementWithKeyValues,
+  key: string,
+): string | null {
+  const keyValue = element.keyValues?.find((kv) => kv?.key === key);
+  // Note: the "values" could be an array with many values.
+  return keyValue?.values?.[0] ?? null;
+}
+
+export function findKeyValueParsed<T = string>(
+  element: ElementWithKeyValues,
+  key: string,
+  parser: (arg0: string) => T,
+): T | null {
+  const keyValue = findKeyValue(element, key);
+  if (keyValue === null) {
+    return keyValue;
+  }
+  return parser(keyValue);
+}
