+++ ui/src/utils/index.ts
@@ -3,6 +3,7 @@ export * from './components';
 export * from './download';
 export * from './enum';
 export * from './errors';
+export * from './findKeyValue';
 export * from './forms';
 export * from './gis';
 export * from './gql';
